+++ matomo_cli/api.py
@@ -17,71 +17,84 @@
 
 """Module providing an API wrapper for Matomo."""
 
+import unicodedata
+import os
 import requests
 import rich_click as click
-import unicodedata
+
+trust_ssl = os.getenv("MATOMO_TRUST_SSL_CERT", "")
+
 
 @click.command()
+@click.option("--url", "-u", envvar="MATOMO_URL", help="URL to Matomo instance")
+@click.option("--token", "-t", envvar="MATOMO_TOKEN", help="Matomo auth token")
 @click.option(
-    "--url", "-u", envvar="MATOMO_URL", help="URL to Matomo instance"
+    "--output_format",
+    "-f",
+    envvar="MATOMO_OUTPUT_FORMAT",
+    default="tsv",
+    help="Format to output (original, json, tsv, and xml supported)",
 )
 @click.option(
-    "--token", "-t", envvar="MATOMO_TOKEN", help="Matomo auth token"
+    "--method",
+    "-m",
+    default="API.getMatomoVersion",
+    help="Method to use (e.g., API.getMatomoVersion)",
 )
 @click.option(
-    "--output_format", "-f", envvar="MATOMO_OUTPUT_FORMAT", default="tsv",
-    help="Format to output (original, json, tsv, and xml supported)"
+    "--id_site",
+    "idSite",
+    "-i",
+    envvar="MATOMO_ID_SITE",
+    default=1,
+    type=int,
+    help="ID site to ask for",
 )
 @click.option(
-    "--method", "-m", default="API.getMatomoVersion",
-    help="Method to use (e.g., API.getMatomoVersion)"
+    "--id_sites",
+    "idSites",
+    "-is",
+    envvar="MATOMO_ID_SITES",
+    default="1",
+    help="Comma-separated list of sites",
+)
+@click.option("--segment_name", "segmentName", "-sn", help="Segment name")
+@click.option(
+    "--period",
+    "-p",
+    envvar="MATOMO_PERIOD",
+    default="day",
+    help="Period to ask for (e.g., day, month, year)",
 )
 @click.option(
-    "--id_site", "idSite", "-i", envvar="MATOMO_ID_SITE", default=1, type=int,
-    help="ID site to ask for"
-)
-@click.option(
-    "--id_sites", "idSites", "-is", envvar="MATOMO_ID_SITES", default="1",
-    help="Comma-separated list of sites"
-)
-@click.option(
-    "--segment_name", "segmentName", "-sn", help="Segment name"
-)
-@click.option(
-    "--period", "-p", envvar="MATOMO_PERIOD", default="day",
-    help="Period to ask for (e.g., day, month, year)"
-)
-@click.option(
-    "--date", "-d", envvar="MATOMO_DATE", default="today",
-    help="Date to use (e.g., today, yesterday, or 2024-10-02)"
+    "--date",
+    "-d",
+    envvar="MATOMO_DATE",
+    default="today",
+    help="Date to use (e.g., today, yesterday, or 2024-10-02)",
 )
 @click.option(
     "--show_columns", "showColumns", "-sc", help="Limit which columns are shown"
 )
+@click.option("--site_name", "siteName", "-sn", help="Site name")
 @click.option(
-    "--site_name", "siteName", "-sn", help="Site name"
+    "--limit",
+    "-l",
+    envvar="MATOMO_LIMIT",
+    type=int,
+    default=1,
+    help="Limit result to this number",
 )
 @click.option(
-    "--api_module", "apiModule", "-am", help="API module"
+    "--offset", "-o", envvar="MATOMO_OFFSET", type=int, default=0, help="Offset result"
 )
 @click.option(
-    "--api_action", "apiAction", "-ac", help="API action"
+    "--extra_params",
+    envvar="MATOMO_EXTRA_PARAMS",
+    default="",
+    help='Comma-separated list of extra options (e.g. "date=today,range=month")',
 )
-@click.option(
-    "--serialize", "-se", type=int, help="Serialize (1 or 0)"
-)
-@click.option(
-    "--limit", "-l", envvar="MATOMO_LIMIT", type=int, default=1,
-    help="Limit result to this number"
-)
-@click.option(
-    "--offset", "-o", envvar="MATOMO_OFFSET", type=int, default=0,
-    help="Offset result"
-)
-
-def api(
-    url, token, output_format, method, **api_params
-) -> None:
+def api(url, token, output_format, extra_params, method, **api_params) -> None:
     """Make API calls to the Matomo instance."""
 
     api_url = url
@@ -92,17 +105,29 @@ def api(
         "token_auth": token,
     }
 
+    # If extra_params are provided, parse them and add to the payload
+    if extra_params:
+        # Split by comma, then by '=' to get key-value pairs
+        extra_dict = dict(item.split("=") for item in extra_params.split(","))
+        # Merge extra options into the payload
+        payload.update(extra_dict)
+
     # Add parameters to payload from api_params
-    payload.update({
-        key: value for key, value in api_params.items() if value is not None
-    })
+    payload.update(
+        {key: value for key, value in api_params.items() if value is not None}
+    )
 
     # Special case handling for method SitesManager.addSite
     if "site_name" in api_params and method == "SitesManager.addSite":
         payload.pop("idSite", None)
 
     # Make API request
-    response = requests.post(api_url, params=payload, timeout=1000)
+    if trust_ssl:
+        response = requests.post(
+            api_url, params=payload, timeout=1000, verify=trust_ssl
+        )
+    else:
+        response = requests.post(api_url, params=payload, timeout=1000)
 
     # Handle the response based on output format
     if response.status_code == 200:
@@ -129,17 +154,20 @@ def print_tsv(tsv_data: str) -> None:
         row = line.split("\t")
         print("\t".join(row))
 
+
 def remove_non_utf8_characters(response):
     """
     Remove non-UTF-8 and other problematic characters from a response.
+    In some cases a call can have chars in first response.
     """
     # Normalize the string to remove special characters
     response = unicodedata.normalize("NFKD", response)
 
     # Remove non-ASCII characters that could be causing issues
-    clean_response = response.encode('ascii', 'ignore').decode('ascii')
+    clean_response = response.encode("ascii", "ignore").decode("ascii")
 
     return clean_response
+
 
 if __name__ == "__main__":
     api()